proto 0.1 sucked, so here's proto 0.2
Why Proto 0.1 sucked:

It was too chatty - expecting eveything to have a REQ/RES pair just
so we could reason about how long it'd been since we heard from a
client was actually very naieve.  It also didn't make account that
MCollective's use allows each agent to be in its own thread and so
attempt to send a reply frame at any point it decides to.  All the
mutexing you need to guard about that makes the whole thing a lot
more lumpy and slower then you really want.  In thinking about ways
to accomodate that, maybe with a connection pool or something I
realised that the 0.1 model was just too dumb.

Why Proto 0.2 is better:

It only sends messages when it needs to.  From connector (MCollective)
side it only sends CONNECT/SUB/UNSUB/PUT/DISCONNECT/PING, and reponds
to PING/PONG.  From the middleware it only sends MESSAGE/ERROR/PING and
responds to PING/PONG.

We can take recent activity as an indication of connection aliveness,
and if we really want to check the other end if still there (we
sent no regular traffic for a while, then a PING/PONG pair is only
a few bytes).  It's actually way closer to STOMP 1.1 with heart-beat
than 0.1 was, and that's not too terrible a thing.

This is working, but without the actual heartbeat logic.  It's
already a load faster than 0.1, as it should be because it only
sends when there's really something to say.  As a side-effect the
middleware doesn't actually have to queue messages anymore, it can
just deliver to the subscribed parties immediately, without worrying
about their phase in the request cycle.

Still need to properly implement heartbeats, and we'll still need
to poll/do mutexes in the mcollective connector as zeromq isn't
thread-safe so the current attempt to disconnect/reconnect in
keepalive_thread causes aborts (send while recv in other thread).

So again 0.19 rather than 0.2.  Keepalive/heartbeat logic soon.