Implement client-side heartbeating and reconnect logic
We have to mutex a little bit here, as zeromq really expects one
socket per thread, and so you need a memory barrier if you want to
access it concurrently as the agent and now keepalive/reconnect
thread will need to.

Also in order to make a new connection with 0MQ you need to make a
new socket, so move the creation of the socket down into connect,
while leaving the one-time setup (loading curve keys) over in
initialize.  We also grow a little more state here, in the
@subscriptions instance variable, to enable us to resubscribe on
reconnect.